pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.6.1",
    "duckduckgo-search>=5.3.0",
    "langchain-community>=0.2.10",
    "langchain-core>=0.2.10",
    "langchain-groq>=0.1.2",
    "langgraph>=0.1.4",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlite-utils>=3.37",
    "streamlit>=1.39.0",
    "typing-extensions>=4.12.2",
]
